fix(macos): zext visibility + pre-existing C++20 issues exposed (#23)
Two originally-reported macOS-clang-only source errors in include/core/uint.h
plus three pre-existing C++20 issues that were masked by the original zext
undeclared-identifier error (which stopped compilation before reaching them).

**include/core/uint.h**

Bug A (uint.h:48): use of undeclared identifier 'zext'
Root cause: zext is defined in core/operators.h at line ~620, but uint.h
is included BEFORE the zext definition is reached (include chain:
io.h -> operators.h:12 -> uint.h). The ch_uint(const ch_bool&) constructor
body is a non-template member of a class template, so clang parses it
eagerly at class definition and finds zext undeclared. GCC defers the
lookup to instantiation, masking the issue. Fix: call node_builder
directly (the same build_operation call that zext would make internally);
node_builder, ch_op, and lnode<T> ARE visible at that point because they
are included earlier in operators.h.

Bug B (uint.h:154): conflicting types for 'make_uint_result'
The friend template at uint.h:154 was redundant: ch_uint is a struct
(default public access), so the constructor explicit ch_uint(lnodeimpl*)
at uint.h:93 is already publicly callable from make_uint_result<Width>
in ch::core namespace. Removing the friend declaration also resolves
the clang conflict between the namespace-scope definition
(operators.h:145) and the in-class friend template.

**include/chlib/switch.h** (pre-existing C++20 issues, now exposed)

switch.h:125: ch::core::ch_uint<1> temp{value} defined a variable of
non-literal type in a constexpr function, which C++20 forbids (C++23
relaxation). Fix: pass the ch_uint<1> as a temporary to zext instead.

switch.h:149: same class of issue, ch_bool is not a literal type in
C++20 (no constexpr constructor). Fix: remove constexpr from
switch_impl; it is only called at simulation/elaboration time, never
from a constant-expression context.

**include/axi4/peripherals/axi_dma.h** (pre-existing C++20 issue, now exposed)

axi_dma.h:63-66: multi-variable auto declarations like
  auto ar_rdy = ..., r_vld = ..., r_rdy_sig = ...;
require all variables to deduce to the same type in C++20 (GCC warned,
clang errors). Fix: split into individual auto declarations.

**Testing**: 133/133 ctest pass (perf tests excluded as pre-existing
unrelated). Full build clean with clang-18 (clang-22 not available
locally; this fix is the prerequisite for the macOS arm64 clang-22
CI matrix in PR #22 to go green).

+++ tests/chlib/test_selector_priority.cpp
@@ -208,7 +208,12 @@ TEST_CASE("SelectorArbiter: priority selector stress tests",
     ch::core::ctx_swap ctx_swapper(ctx.get());
 
     SECTION("Maximum width priority selector") {
-        ch_uint<64> request(0xFFFFFFFFFFFFFFFFUL); // All 64 bits set
+        // Explicit uint64_t cast: on macOS LP64, `unsigned long` is
+        // 64-bit, same as uint64_t, so `0xFFFFFFFFFFFFFFFFUL` is
+        // ambiguous between ch_literal_runtime(uint64_t) and
+        // ch_literal_runtime(int64_t) constructors. Disambiguating
+        // with an explicit type is portable across macOS/Linux.
+        ch_uint<64> request(uint64_t(0xFFFFFFFFFFFFFFFFUL)); // All 64 bits set
         PrioritySelectorResult<64> result = priority_selector<64>(request);
 
         ch::Simulator sim(ctx.get());
